requirements.txt
google-generativeai==0.7.2
tabulate==0.9.0
python-dotenv==1.0.1
